separate out span code and add better comments
