+++ crates/oxc_linter_plugin/src/lib.rs
@@ -5,4 +5,6 @@ mod plugin;
 #[cfg(test)]
 mod raw_diagnostic;
 #[cfg(test)]
+mod spans;
+#[cfg(test)]
 mod test;
